LPS-172038 portal-search-web: Move and create base class
https://issues.liferay.com/browse/LPS-172038
