+++ modules/apps/portal-search/portal-search-web/src/test/java/com/liferay/portal/search/web/internal/facet/display/context/UserSearchFacetDisplayContextTest.java
@@ -18,9 +18,9 @@ import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.search.facet.Facet;
 import com.liferay.portal.kernel.search.facet.collector.FacetCollector;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.search.web.internal.BaseFacetDisplayContextTestCase;
 import com.liferay.portal.search.web.internal.facet.display.context.builder.UserSearchFacetDisplayContextBuilder;
 import com.liferay.portal.search.web.internal.user.facet.configuration.UserFacetPortletInstanceConfiguration;
-import com.liferay.portal.search.web.internal.util.FacetDisplayContextTextUtil;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
 
 import java.util.List;
@@ -106,7 +106,7 @@ public class UserSearchFacetDisplayContextTest {
 
 		int count = RandomTestUtil.randomInt();
 
-		FacetDisplayContextTextUtil.setUpTermCollector(
+		BaseFacetDisplayContextTestCase.setUpTermCollector(
 			_facetCollector, userName, count);
 
 		String paramValue = "";
@@ -140,7 +140,7 @@ public class UserSearchFacetDisplayContextTest {
 
 		int count = RandomTestUtil.randomInt();
 
-		FacetDisplayContextTextUtil.setUpTermCollector(
+		BaseFacetDisplayContextTestCase.setUpTermCollector(
 			_facetCollector, userName, count);
 
 		String paramValue = userName;
@@ -172,16 +172,16 @@ public class UserSearchFacetDisplayContextTest {
 	public void testOrderByTermFrequencyAscending() throws Exception {
 		String[] userNames = {"charlie", "delta", "bravo", "alpha"};
 
-		FacetDisplayContextTextUtil.setUpTermCollectors(
+		BaseFacetDisplayContextTestCase.setUpTermCollectors(
 			_facetCollector,
-			FacetDisplayContextTextUtil.getTermCollectors(
+			BaseFacetDisplayContextTestCase.getTermCollectors(
 				userNames, new int[] {6, 5, 5, 4}));
 
 		UserSearchFacetDisplayContext userSearchFacetDisplayContext =
 			_createDisplayContext(StringPool.BLANK, "count:asc");
 
 		String nameFrequencyString =
-			FacetDisplayContextTextUtil.buildNameFrequencyString(
+			BaseFacetDisplayContextTestCase.buildNameFrequencyString(
 				userSearchFacetDisplayContext.getBucketDisplayContexts());
 
 		Assert.assertEquals(
@@ -192,16 +192,16 @@ public class UserSearchFacetDisplayContextTest {
 	public void testOrderByTermFrequencyDescending() throws Exception {
 		String[] userNames = {"alpha", "delta", "bravo", "charlie"};
 
-		FacetDisplayContextTextUtil.setUpTermCollectors(
+		BaseFacetDisplayContextTestCase.setUpTermCollectors(
 			_facetCollector,
-			FacetDisplayContextTextUtil.getTermCollectors(
+			BaseFacetDisplayContextTestCase.getTermCollectors(
 				userNames, new int[] {4, 5, 5, 6}));
 
 		UserSearchFacetDisplayContext userSearchFacetDisplayContext =
 			_createDisplayContext(StringPool.BLANK, "count:desc");
 
 		String nameFrequencyString =
-			FacetDisplayContextTextUtil.buildNameFrequencyString(
+			BaseFacetDisplayContextTestCase.buildNameFrequencyString(
 				userSearchFacetDisplayContext.getBucketDisplayContexts());
 
 		Assert.assertEquals(
@@ -210,9 +210,9 @@ public class UserSearchFacetDisplayContextTest {
 
 	@Test
 	public void testOrderByTermValueAscending() throws Exception {
-		FacetDisplayContextTextUtil.setUpTermCollectors(
+		BaseFacetDisplayContextTestCase.setUpTermCollectors(
 			_facetCollector,
-			FacetDisplayContextTextUtil.getTermCollectors(
+			BaseFacetDisplayContextTestCase.getTermCollectors(
 				new String[] {"bravo", "alpha", "bravo", "charlie"},
 				new int[] {3, 4, 5, 6}));
 
@@ -220,7 +220,7 @@ public class UserSearchFacetDisplayContextTest {
 			_createDisplayContext(StringPool.BLANK, "key:asc");
 
 		String nameFrequencyString =
-			FacetDisplayContextTextUtil.buildNameFrequencyString(
+			BaseFacetDisplayContextTestCase.buildNameFrequencyString(
 				userSearchFacetDisplayContext.getBucketDisplayContexts());
 
 		Assert.assertEquals(
@@ -229,9 +229,9 @@ public class UserSearchFacetDisplayContextTest {
 
 	@Test
 	public void testOrderByTermValueDescending() throws Exception {
-		FacetDisplayContextTextUtil.setUpTermCollectors(
+		BaseFacetDisplayContextTestCase.setUpTermCollectors(
 			_facetCollector,
-			FacetDisplayContextTextUtil.getTermCollectors(
+			BaseFacetDisplayContextTestCase.getTermCollectors(
 				new String[] {"bravo", "alpha", "bravo", "charlie"},
 				new int[] {3, 4, 5, 6}));
 
@@ -239,7 +239,7 @@ public class UserSearchFacetDisplayContextTest {
 			_createDisplayContext(StringPool.BLANK, "key:desc");
 
 		String nameFrequencyString =
-			FacetDisplayContextTextUtil.buildNameFrequencyString(
+			BaseFacetDisplayContextTestCase.buildNameFrequencyString(
 				userSearchFacetDisplayContext.getBucketDisplayContexts());
 
 		Assert.assertEquals(
@@ -260,7 +260,7 @@ public class UserSearchFacetDisplayContextTest {
 		UserSearchFacetDisplayContextBuilder
 			userSearchFacetDisplayContextBuilder =
 				new UserSearchFacetDisplayContextBuilder(
-					FacetDisplayContextTextUtil.getRenderRequest(
+					BaseFacetDisplayContextTestCase.getRenderRequest(
 						UserFacetPortletInstanceConfiguration.class));
 
 		userSearchFacetDisplayContextBuilder.setFacet(_facet);
